requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
python-dotenv==1.0.1
pypdf==4.0.1
numpy==1.26.4
reportlab==4.1.0
vercel-edge==0.1.2
typing-extensions>=4.5.0 